move stuff around, clean up structure
