+++ frontend/src/Components/UI/card.tsx
@@ -1,6 +1,6 @@
 import * as React from "react"
 
-import { cn } from "../../lib/utils"
+import { cn } from "../utils.ts"
 
 const Card = React.forwardRef<
   HTMLDivElement,
