+++ frontend/src/Components/UI/dropdownMenu.tsx
@@ -2,7 +2,7 @@ import * as React from "react"
 import * as DropdownMenuPrimitive from "@radix-ui/react-dropdown-menu"
 import { Check, ChevronRight, Circle } from "lucide-react"
 
-import { cn } from "../../lib/utils"
+import { cn } from "../utils.ts"
 
 const DropdownMenu = DropdownMenuPrimitive.Root
