+++ frontend/src/Components/modeToggle.tsx
@@ -1,13 +1,13 @@
 import { Moon, Sun } from "lucide-react"
 
-import { Button } from "../components/ui/button"
+import { Button } from "./UI/button"
 import {
   DropdownMenu,
   DropdownMenuContent,
   DropdownMenuItem,
   DropdownMenuTrigger,
-} from "../components/ui/dropdown-menu"
-import { useTheme } from "../components/theme-provider"
+} from "./UI/dropdownMenu.tsx"
+import { useTheme } from "./themeProvider.tsx"
 
 export function ModeToggle() {
   const { setTheme } = useTheme()
