+++ frontend/src/WelcomePage.tsx
@@ -1,5 +1,5 @@
 import { FormEvent, useState } from "react";
-import { Input } from "./components/ui/input";
+import { Input } from "./Components/UI/input";
 
 
 export const WelcomePage = () => {
